backend/requirements.txt
# ===========================================
# BikeFit Pro - Backend Dependencies
# ===========================================
# Python 3.11+
# ===========================================

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0
asyncpg>=0.29.0
greenlet>=3.0.0

# Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Pose Estimation (YOLOv8)
ultralytics>=8.1.0
opencv-python>=4.9.0
numpy>=1.26.0

# PDF Generation
reportlab>=4.1.0
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
